requirements.txt
Django>=1.9.6
django-datetime-widget>=0.9.3
django-multiupload>=0.5.1
django-summernote>=0.8.5
docker-py>=1.8.1
psutil>=4.2.0
pytz>=2016.4
requests>=2.10.0
six>=1.10.0
websocket-client>=0.37.0
whitenoise>=3.1
